done with fetching data from server

+++ app/src/main/java/com/example/movielistappnew/Model/Movie.java
@@ -7,58 +7,46 @@ import java.util.List;
 
 public class Movie {
 
+
     @SerializedName("adult")
     @Expose
     private Boolean adult;
-
     @SerializedName("backdrop_path")
     @Expose
     private String backdropPath;
-
     @SerializedName("genre_ids")
     @Expose
     private List<Integer> genreIds = null;
-
     @SerializedName("id")
     @Expose
     private Integer id;
-
     @SerializedName("original_language")
     @Expose
     private String originalLanguage;
-
     @SerializedName("original_title")
     @Expose
     private String originalTitle;
-
     @SerializedName("overview")
     @Expose
     private String overview;
-
     @SerializedName("popularity")
     @Expose
     private Double popularity;
-
     @SerializedName("poster_path")
     @Expose
     private String posterPath;
-
     @SerializedName("release_date")
     @Expose
     private String releaseDate;
-
     @SerializedName("title")
     @Expose
     private String title;
-
     @SerializedName("video")
     @Expose
     private Boolean video;
-
     @SerializedName("vote_average")
     @Expose
-    private Integer voteAverage;
-
+    private Double voteAverage;
     @SerializedName("vote_count")
     @Expose
     private Integer voteCount;
@@ -159,11 +147,11 @@ public class Movie {
         this.video = video;
     }
 
-    public Integer getVoteAverage() {
+    public Double getVoteAverage() {
         return voteAverage;
     }
 
-    public void setVoteAverage(Integer voteAverage) {
+    public void setVoteAverage(Double voteAverage) {
         this.voteAverage = voteAverage;
     }
